extract without specobj
