+++ pypeit/core/skysub.py
@@ -3,21 +3,26 @@
 .. include common links, assuming primary doc root is up one directory
 .. include:: ../include/links.rst
 """
-import numpy as np
+from pathlib import Path
 
+from IPython import embed
+import numpy as np
+import matplotlib.pyplot as plt
 import scipy.ndimage
 import scipy.special
 from scipy.interpolate import RegularGridInterpolator
 
-import matplotlib.pyplot as plt
-
-from IPython import embed
-
-from pypeit.core import basis, pixels, extract
-from pypeit.core import fitting
-from pypeit.core import procimg
-from pypeit import log, utils, bspline, slittrace
+from pypeit import bspline
+from pypeit import log
 from pypeit import PypeItError
+from pypeit import slittrace
+from pypeit import utils
+from pypeit.core import basis
+from pypeit.core import extract
+from pypeit.core import fitting
+from pypeit.core import pixels
+from pypeit.core import procimg
+from pypeit.core import spatialprofile
 from pypeit.display import display
 
 
@@ -862,9 +867,38 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, thismask, 
                     log.info("------------------------------------------------------------------------------------------------------------")
 
                     # TODO -- Use extract_specobj_boxcar to avoid code duplication
-                    extract.extract_boxcar(sciimg-skyimage, modelivar, outmask, waveimg, skyimage,
-                                           sobjs[iobj], fwhmimg=fwhmimg, base_var=base_var, count_scale=count_scale,
-                                           noise_floor=adderr)
+
+                    if sobjs[iobj].NAME == 'SPAT0714-SLIT0715-DET01':
+                        ofile = f'{sobjs[iobj].NAME}_before_boxcar_{iiter}.fits'
+                        if not Path(ofile).is_file():
+                            sobjs[iobj].to_file(ofile)
+#                            ofile_inputs = f'{sobjs[iobj].NAME}_boxcar_inputs_{iiter}.npz'
+#                            np.savez_compressed(ofile_inputs,
+#                                sciimg=sciimg, modelivar=modelivar, outmask=outmask,
+#                                waveimg=waveimg, skyimage=skyimage, fwhmimg=fwhmimg,
+#                                base_var=base_var, count_scale=count_scale, noise_floor=adderr
+#                            )
+
+                    if sobjs[iobj].trace_spec is None:
+                        sobjs[iobj].trace_spec = np.arange(sciimg.shape[0])
+                    (
+                        sobjs[iobj].BOX_WAVE, sobjs[iobj].BOX_COUNTS, sobjs[iobj].BOX_COUNTS_IVAR,
+                        sobjs[iobj].BOX_COUNTS_SIG, sobjs[iobj].BOX_COUNTS_NIVAR,
+                        sobjs[iobj].BOX_MASK, sobjs[iobj].BOX_FWHM, sobjs[iobj].BOX_FLAT,
+                        sobjs[iobj].BOX_COUNTS_SKY, sobjs[iobj].BOX_COUNTS_SIG_DET,
+                        sobjs[iobj].BOX_NPIX
+                    ) = extract.extract_boxcar(
+                        sobjs[iobj].BOX_R_PIX, sobjs[iobj].TRACE_SPAT, sciimg-skyimage, modelivar,
+                        outmask, waveimg, skyimage, sobjs[iobj], fwhmimg=fwhmimg,
+                        base_var=base_var, count_scale=count_scale, noise_floor=adderr,
+                        trace_spec=sobjs[iobj].trace_spec
+                    )
+
+                    if sobjs[iobj].NAME == 'SPAT0714-SLIT0715-DET01':
+                        ofile = f'{sobjs[iobj].NAME}_after_boxcar_{iiter}.fits'
+                        if not Path(ofile).is_file():
+                            sobjs[iobj].to_file(ofile)
+
                     flux = sobjs[iobj].BOX_COUNTS
                     fluxivar = sobjs[iobj].BOX_COUNTS_IVAR * sobjs[iobj].BOX_MASK
                     wave = sobjs[iobj].BOX_WAVE
@@ -874,17 +908,68 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, thismask, 
                     trace = sobjs[iobj].TRACE_SPAT[:, None]
                     objmask = ((spat_img >= (trace - 2.0 * sobjs[iobj].BOX_R_PIX)) & (spat_img <= (trace + 2.0 * sobjs[iobj].BOX_R_PIX)))
                     # Boxcar
-                    extract.extract_boxcar(sciimg-skyimage, modelivar, (outmask & objmask), waveimg,
-                                           skyimage, sobjs[iobj], fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var,
-                                           count_scale=count_scale, noise_floor=adderr)
+
+                    if sobjs[iobj].NAME == 'SPAT0714-SLIT0715-DET01':
+                        ofile = f'{sobjs[iobj].NAME}_before_boxcar_{iiter}.fits'
+                        if not Path(ofile).is_file():
+                            sobjs[iobj].to_file(ofile)
+#                            ofile_inputs = f'{sobjs[iobj].NAME}_boxcar_inputs_{iiter}.npz'
+#                            np.savez_compressed(ofile_inputs,
+#                                sciimg=sciimg, modelivar=modelivar, outmask=outmask & objmask,
+#                                waveimg=waveimg, skyimage=skyimage, fwhmimg=fwhmimg,
+#                                flatimg=flatimg, base_var=base_var, count_scale=count_scale,
+#                                noise_floor=adderr
+#                            )
+
+                    (
+                        sobjs[iobj].BOX_WAVE, sobjs[iobj].BOX_COUNTS, sobjs[iobj].BOX_COUNTS_IVAR,
+                        sobjs[iobj].BOX_COUNTS_SIG, sobjs[iobj].BOX_COUNTS_NIVAR,
+                        sobjs[iobj].BOX_MASK, sobjs[iobj].BOX_FWHM, sobjs[iobj].BOX_FLAT,
+                        sobjs[iobj].BOX_COUNTS_SKY, sobjs[iobj].BOX_COUNTS_SIG_DET,
+                        sobjs[iobj].BOX_NPIX
+                    ) = extract.extract_boxcar(
+                        sobjs[iobj].BOX_R_PIX, sobjs[iobj].TRACE_SPAT, sciimg-skyimage, modelivar,
+                        (outmask & objmask), waveimg, skyimage, fwhmimg=fwhmimg, flatimg=flatimg,
+                        base_var=base_var, count_scale=count_scale, noise_floor=adderr,
+                        trace_spec=sobjs[iobj].trace_spec
+                    )
+
+                    if sobjs[iobj].NAME == 'SPAT0714-SLIT0715-DET01':
+                        ofile = f'{sobjs[iobj].NAME}_after_boxcar_{iiter}.fits'
+                        if not Path(ofile).is_file():
+                            sobjs[iobj].to_file(ofile)
+#                            ofile_inputs = f'{sobjs[iobj].NAME}_optimal_inputs_{iiter}.npz'
+#                            np.savez_compressed(ofile_inputs,
+#                                sciimg=sciimg, modelivar=modelivar, outmask=outmask & objmask,
+#                                waveimg=waveimg, skyimage=skyimage, thismask=thismask,
+#                                last_profile=last_profile, min_frac_use=min_frac_use,
+#                                fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var,
+#                                count_scale=count_scale, noise_floor=adderr
+#                            )
+
                     # Optimal
-                    extract.extract_optimal(sciimg-skyimage, modelivar, (outmask & objmask), waveimg,
-                                            skyimage, thismask, last_profile, sobjs[iobj], min_frac_use=min_frac_use,
-                                            fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var, count_scale=count_scale,
-                                            noise_floor=adderr)
+                    (
+                        sobjs[iobj].OPT_WAVE, sobjs[iobj].OPT_COUNTS, sobjs[iobj].OPT_COUNTS_IVAR,
+                        sobjs[iobj].OPT_COUNTS_SIG, sobjs[iobj].OPT_COUNTS_NIVAR,
+                        sobjs[iobj].OPT_MASK, sobjs[iobj].OPT_FWHM, sobjs[iobj].OPT_FLAT,
+                        sobjs[iobj].OPT_COUNTS_SKY, sobjs[iobj].OPT_COUNTS_SIG_DET,
+                        sobjs[iobj].OPT_FRAC_USE, sobjs[iobj].OPT_CHI2
+                    ) = extract.extract_optimal(
+                        sciimg-skyimage, modelivar, (outmask & objmask), waveimg, skyimage,
+                        thismask, last_profile, min_frac_use=min_frac_use, fwhmimg=fwhmimg,
+                        flatimg=flatimg, base_var=base_var, count_scale=count_scale,
+                        noise_floor=adderr, box_radius=sobjs[iobj].BOX_R_PIX,
+                        trace_spec=sobjs[iobj].trace_spec, trace_spat=sobjs[iobj].TRACE_SPAT
+                    )
+                    
+                    if sobjs[iobj].NAME == 'SPAT0714-SLIT0715-DET01':
+                        ofile = f'{sobjs[iobj].NAME}_after_optimal_{iiter}.fits'
+                        if not Path(ofile).is_file():
+                            sobjs[iobj].to_file(ofile)
+
                     # If the extraction is bad do not update
                     if sobjs[iobj].OPT_MASK is not None:
-                        # if there is only one good pixel `extract.fit_profile` fails
+                        # if there is only one good pixel `spatialprofile.fit_profile` fails
                         if np.sum(sobjs[iobj].OPT_MASK) > extract_good_frac * nspec:
                             flux = sobjs[iobj].OPT_COUNTS
                             fluxivar = sobjs[iobj].OPT_COUNTS_IVAR*sobjs[iobj].OPT_MASK
@@ -894,7 +979,7 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, thismask, 
                 if wave.any():
                     sign = sobjs[iobj].sign
                     # TODO This is "sticky" masking. Do we want it to be?
-                    profile_model, trace_new, fwhmfit, med_sn2 = extract.fit_profile(
+                    profile_model, trace_new, fwhmfit, med_sn2 = spatialprofile.fit_profile(
                         sign*img_minsky[ipix], (modelivar * outmask)[ipix],waveimg[ipix], thismask[ipix], spat_pix[ipix], sobjs[iobj].TRACE_SPAT,
                         wave, sign*flux, fluxivar, inmask = outmask[ipix],
                         thisfwhm=sobjs[iobj].FWHM, prof_nsigma=sobjs[iobj].prof_nsigma, sn_gauss=sn_gauss, gauss=force_gauss, obj_string=obj_string,
@@ -1011,15 +1096,39 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, thismask, 
             trace = sobjs[iobj].TRACE_SPAT[:, None]
             # Optimal
             objmask = ((spat_img >= (trace - 2.0 * sobjs[iobj].BOX_R_PIX)) & (spat_img <= (trace + 2.0 * sobjs[iobj].BOX_R_PIX)))
-            extract.extract_optimal(sciimg-skyimage, modelivar * thismask, (outmask_extract & objmask),
-                                    waveimg, extract_sky, thismask, this_profile, sobjs[iobj], min_frac_use=min_frac_use,
-                                    fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var, count_scale=count_scale,
-                                    noise_floor=adderr)
+            (
+                sobjs[iobj].OPT_WAVE, sobjs[iobj].OPT_COUNTS, sobjs[iobj].OPT_COUNTS_IVAR,
+                sobjs[iobj].OPT_COUNTS_SIG, sobjs[iobj].OPT_COUNTS_NIVAR,
+                sobjs[iobj].OPT_MASK, sobjs[iobj].OPT_FWHM, sobjs[iobj].OPT_FLAT,
+                sobjs[iobj].OPT_COUNTS_SKY, sobjs[iobj].OPT_COUNTS_SIG_DET,
+                sobjs[iobj].OPT_FRAC_USE, sobjs[iobj].OPT_CHI2
+            ) = extract.extract_optimal(
+                sciimg-skyimage, modelivar*thismask, (outmask_extract & objmask), waveimg,
+                extract_sky, thismask, this_profile, min_frac_use=min_frac_use, fwhmimg=fwhmimg,
+                flatimg=flatimg, base_var=base_var, count_scale=count_scale, noise_floor=adderr,
+                box_radius=sobjs[iobj].BOX_R_PIX, trace_spec=sobjs[iobj].trace_spec,
+                trace_spat=sobjs[iobj].TRACE_SPAT
+            )
+#            extract.extract_optimal(sciimg-skyimage, modelivar * thismask, (outmask_extract & objmask),
+#                                    waveimg, extract_sky, thismask, this_profile, sobjs[iobj], min_frac_use=min_frac_use,
+#                                    fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var, count_scale=count_scale,
+#                                    noise_floor=adderr)
             # Boxcar
-            extract.extract_boxcar(sciimg-skyimage, modelivar*thismask, (outmask_extract & objmask),
-                                   waveimg, extract_sky, sobjs[iobj],
-                                   fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var,
-                                   count_scale=count_scale, noise_floor=adderr)
+            (
+                sobjs[iobj].BOX_WAVE, sobjs[iobj].BOX_COUNTS, sobjs[iobj].BOX_COUNTS_IVAR,
+                sobjs[iobj].BOX_COUNTS_SIG, sobjs[iobj].BOX_COUNTS_NIVAR, sobjs[iobj].BOX_MASK,
+                sobjs[iobj].BOX_FWHM, sobjs[iobj].BOX_FLAT, sobjs[iobj].BOX_COUNTS_SKY,
+                sobjs[iobj].BOX_COUNTS_SIG_DET, sobjs[iobj].BOX_NPIX
+            ) = extract.extract_boxcar(
+                sobjs[iobj].BOX_R_PIX, sobjs[iobj].TRACE_SPAT, sciimg-skyimage, modelivar*thismask,
+                (outmask_extract & objmask), waveimg, extract_sky, fwhmimg=fwhmimg, flatimg=flatimg,
+                base_var=base_var, count_scale=count_scale, noise_floor=adderr,
+                trace_spec=sobjs[iobj].trace_spec
+            )
+#            extract.extract_boxcar(sciimg-skyimage, modelivar*thismask, (outmask_extract & objmask),
+#                                   waveimg, extract_sky, sobjs[iobj],
+#                                   fwhmimg=fwhmimg, flatimg=flatimg, base_var=base_var,
+#                                   count_scale=count_scale, noise_floor=adderr)
             sobjs[iobj].min_spat = min_spat
             sobjs[iobj].max_spat = max_spat
