+++ pypeit/scripts/multislit_flexure.py
@@ -4,19 +4,7 @@ Script to measure and correct for flexure in multi-slit data.
 .. include common links, assuming primary doc root is up one directory
 .. include:: ../include/links.rst
 """
-import os
-
-from IPython import embed
-
-import numpy as np
-
-from pypeit import log
-from pypeit import inputfiles
-from pypeit.spectrographs.util import load_spectrograph
-from pypeit.par import pypeitpar
-from pypeit.core import flexure
 from pypeit.scripts import scriptbase
-
 
 # TODO: Maybe not a good idea to name this script the same as the
 # flexure.MultiSlitFlexure class, but it is technically okay...
@@ -49,6 +37,13 @@ class MultiSlitFlexure(scriptbase.ScriptBase):
     def main(cls, args):
 
         from astropy.io import fits
+        from IPython import embed
+
+        from pypeit import log
+        from pypeit import inputfiles
+        from pypeit.spectrographs.util import load_spectrograph
+        from pypeit.par import pypeitpar
+        from pypeit.core import flexure
 
         # Initialize the log
         cls.init_log(args)
